resolve generic and native touchable circ deps
