+++ src/components/touchables/TouchableNativeFeedbackProps.tsx
@@ -1,0 +1,12 @@
+import { TouchableNativeFeedbackProps as RNTouchableNativeFeedbackProps } from 'react-native';
+import { GenericTouchableProps } from './GenericTouchableProps';
+
+export type TouchableNativeFeedbackExtraProps = {
+  borderless?: boolean;
+  rippleColor?: number | null;
+  rippleRadius?: number | null;
+  foreground?: boolean;
+};
+
+export type TouchableNativeFeedbackProps = RNTouchableNativeFeedbackProps &
+  GenericTouchableProps;
